+++ bin/studymanager/cs_studymanager_study.py
@@ -181,6 +181,10 @@ class Case(object):
         self.estim_wtime = data['estim_wtime']
         self.depends     = data['depends']
 
+        self.parametric  = data['parametric']
+        self.notebook    = data['notebook']
+        self.kw_args     = data['kw_args']
+
         self.is_compiled = "not done"
         self.is_run      = "not done"
         self.is_time     = None
@@ -402,6 +406,17 @@ class Case(object):
         self.run_dir = run_dir
 
         run_cmd = enquote_arg(self.exe) + " run --id=" + enquote_arg(self.run_id)
+
+        if self.notebook:
+            run_cmd += ' --notebook-args ' + self.notebook
+
+        if self.parametric:
+            run_cmd += ' --parametric-args ' + '"' + self.parametric + '"'
+
+        if self.kw_args:
+            if self.kw_args.find(" ") < 0:
+                self.kw_args += " "  # workaround for arg-parser issue
+            run_cmd += ' --kw-args ' + '"' + self.kw_args + '"'
 
         n_procs = self.__data['n_procs']
         if n_procs: